products_api/requirements.txt
black
django
django-mptt
djangorestframework
drf-spectacular
flake8
pytest
pytest-django
python-dotenv
pytest-cov
pytest-factoryboy